+++ contracts/vending-minter/src/testing/tests.rs
@@ -1,2 +1,6 @@
-mod integration_tests;
-mod test_splits;
+mod address_limit;
+mod happy_unhappy;
+mod mint_and_burn;
+mod splits;
+mod trading_time;
+mod whitelist;
